Added: `AccountAndBalanceOperationsTest`
